refactor: migrate SwiftData models to versioned schema
Wrap all @Model classes inside BangumiSchemaV2 (VersionedSchema),
keeping only stored properties and initializers in the schema.
Computed properties and methods are moved to extensions
on typealiases in the original model files.

This follows the pattern established in KMReader, enabling
proper schema versioning and migrations for future changes.

Changes:
- Add BangumiSchemaV2 enum with all 13 @Model classes
- Add BangumiMigrationPlan for future migration stages
- Move computed vars/funcs out of schema into extensions
- Update MainApp and Mock to use versioned schema
- No model property changes (non-breaking)
