+++ App/Models/Schema/BangumiMigrationPlan.swift
@@ -1,0 +1,11 @@
+import SwiftData
+
+enum BangumiMigrationPlan: SchemaMigrationPlan {
+    static var schemas: [any VersionedSchema.Type] {
+        [BangumiSchemaV2.self]
+    }
+
+    static var stages: [MigrationStage] {
+        []
+    }
+}
