+++ App/Models/Trending+Model.swift
@@ -1,19 +1,4 @@
 import Foundation
-import OSLog
 import SwiftData
-import SwiftUI
 
-typealias TrendingSubject = TrendingSubjectV1
-
-@Model
-final class TrendingSubjectV1 {
-  @Attribute(.unique)
-  var type: Int
-
-  var items: [TrendingSubjectDTO]
-
-  init(type: Int, items: [TrendingSubjectDTO]) {
-    self.type = type
-    self.items = items
-  }
-}
+typealias TrendingSubject = BangumiSchemaV2.TrendingSubjectV1
